refresh add file

+++ libaresq/record.h
@@ -68,8 +68,8 @@ public:
 
 	// uint24(_data + 14)
 	inline uint32_t size24() const { return p2l24(_data + 14); }
-	inline void size24(size_t size) { l2p24(size, _data + 14); }
-	inline bool sizeChanged(size_t size) { return (size & 0xffffff) == size24(); }
+	inline void size24(uint64_t size) { l2p24((uint32_t)size, _data + 14); }
+	inline bool sizeChanged(uint64_t size) { return (size & 0xffffff) == size24(); }
 
 	// for dir item (parent/hist, sub)
 	// uint24(_data + 11): isdel() ? parent : hist
